Moved logic for grouping by dates into utils class, updated comment.
