+++ C#/TollCalculator/TollCalculator.Lib/TollCalculator.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using TollCalculator.Lib.Models;
+using TollCalculator.Lib.Utils;
 
 namespace TollCalculator.Lib
 { 
@@ -9,29 +9,8 @@ namespace TollCalculator.Lib
     {
         public static int GetTollFee(VehicleType vehicleType, DateTime[] dates)
         {
-            var datesGroupedByDay = GetDatesGroupedByDay(dates);
+            var datesGroupedByDay = Groups.GetDatesGroupedByDay(dates);
             return datesGroupedByDay.Sum(g => GetDailyFee(g, vehicleType));
-        }
-
-        private static List<SameDayGroup> GetDatesGroupedByDay(DateTime[] dates)
-        {
-            var groups = new List<SameDayGroup>();
-
-            foreach (var date in dates)
-            {
-                var existingGroup = groups.Find(g => g.IsSameDay(date));
-                var hasExisting = existingGroup != null;
-
-                if (hasExisting)
-                    existingGroup.AddDate(date);
-                else
-                {
-                    var newGroup = new SameDayGroup(date);
-                    groups.Add(newGroup);
-                }
-            }
-
-            return groups;
         }
 
         private static int GetDailyFee(SameDayGroup group, VehicleType vehicleType)
@@ -42,7 +21,7 @@ namespace TollCalculator.Lib
                 CheckIsTollFreeDay(group);
                 CheckIsTollFreeHoliday(group);
 
-                var datesGroupedByHour = GetDatesGroupedByHour(group.Dates);
+                var datesGroupedByHour = Groups.GetDatesGroupedByHour(group.Dates);
                 var hourGroupsTotal = datesGroupedByHour.Sum(GetHourlyFee);
 
                 return GetFeeInsideDailyLimit(hourGroupsTotal);
@@ -75,27 +54,6 @@ namespace TollCalculator.Lib
 
             if (isHoliday)
                 throw new FeeExemptedException();
-        }
-
-        private static List<SameHourGroup> GetDatesGroupedByHour(List<DateTime> dates)
-        {
-            var groups = new List<SameHourGroup>();
-
-            foreach (var date in dates)
-            {
-                var existingGroup = groups.Find(g => g.IsSameHour(date));
-                var hasExisting = existingGroup != null;
-
-                if (hasExisting)
-                    existingGroup.AddDate(date);
-                else
-                {
-                    var newGroup = new SameHourGroup(date);
-                    groups.Add(newGroup);
-                }
-            }
-
-            return groups;
         }
 
         private static int GetHourlyFee(SameHourGroup group)
@@ -141,10 +99,10 @@ namespace TollCalculator.Lib
              * 1) Implement logic that can calculate the dynamic holidays for a year based on the formulas
              * available for Swedish public holidays (e.g. https://sv.wikipedia.org/wiki/Helgdagar_i_Sverige)
              * 2) Start using a third-party library that solves this, e.g. https://www.nuget.org/packages/Nager.Date/,
-             * However, such a library should be vetted in the sense if the license complies with ours and if we can trust
-             * the code or not.
+             * however, such a library should be vetted on the criteria 1) if the license complies with ours, 2) if
+             * we can trust the source of the library and 3) if we can trust the quality of the code
              *
-             * Once dynamic holdays are added to the code, the unit tests should also reflect test cases with dynamic holidays.
+             * Once dynamic holidays are added to the code, the unit tests should also reflect test cases with dynamic holidays.
              */
 
             return SwedishHolidays.StaticPublicHolidays.Any(holiday => holiday.IsEqualToDate(group.Month, group.Day));
